requirements.txt
numpy
matplotlib
scipy
xarray
dask
zarr
netCDF4
torch
seaborn
cartopy
pyyaml
pandas
jupyter
notebook
hydra-core